Add files via upload
On account of my struggles with PyBank I wasn't able to tackle this to the depth I needed to but I have a strong grasp of things now because of PyBank. I also have a good feeling that Pandas holds all of the more elegant solutions and removes some of the caveman dentistry.