Add files via upload
I struggled mightily with this project. It took me from Sunday to Thursday night. I finally consulted a friend on Friday and he looked over my shoulder as I figured it out. It was tough though and after struggling with the csvreader object now I understand how to use it. But it took awhile.  Because of this I wasn't able to spend much time on the 2nd part of the homework but I feel that pandas will provide a more elegant solution to these challenges. -Jeff